Return -1 when display big str

+++ ft_printf_c.c
@@ -6,13 +6,13 @@
 /*   By: iyamada <iyamada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/14 06:26:53 by iyamada           #+#    #+#             */
-/*   Updated: 2021/12/06 12:19:20 by iyamada          ###   ########.fr       */
+/*   Updated: 2021/12/07 02:17:53 by iyamada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_printf_c(va_list *ap, int write_len)
+size_t	ft_printf_c(va_list *ap, size_t write_len)
 {
 	ft_putchar(va_arg(*ap, int));
 	return (++write_len);
