+++ ft_printf_percent.c
@@ -6,13 +6,13 @@
 /*   By: iyamada <iyamada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/14 14:09:18 by iyamada           #+#    #+#             */
-/*   Updated: 2021/12/06 12:18:33 by iyamada          ###   ########.fr       */
+/*   Updated: 2021/12/07 02:20:44 by iyamada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_printf_percent(int write_len)
+size_t	ft_printf_percent(size_t write_len)
 {
 	ft_putchar('%');
 	return (++write_len);
